完成文章发布

+++ src/wiki/laona/domain/Article.java
@@ -16,7 +16,7 @@ public class Article {
     private Integer articleId;
     private String articleTitle;
     private String articleContext;
-    private Integer articleTime;
+    private Long articleTime;
     private String articlePic;
     private String articleDesc;
     /**
